ChatShield/requirements.txt
streamlit>=1.32.0
openai>=1.12.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
faiss-cpu>=1.7.4
spacy>=3.7.0
nltk>=3.8.1
pdfplumber>=0.10.3
python-docx>=1.1.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
tiktoken>=0.5.2
numpy>=1.24.0 